# artifacts/api-server/src/routes/searches.ts
import { Router, type IRouter } from "express";
import { db, missingPersonsTable, searchesTable, alertsTable } from "@workspace/db";
import { eq, desc } from "drizzle-orm";
import { PerformSearchBody } from "@workspace/api-zod";

const router: IRouter = Router();

function cosineSimilarity(a: number[], b: number[]): number {
  let dot = 0;
  let normA = 0;
  let normB = 0;
  for (let i = 0; i < a.length; i++) {
    dot += a[i] * b[i];
    normA += a[i] * a[i];
    normB += b[i] * b[i];
  }
  if (normA === 0 || normB === 0) return 0;
  return dot / (Math.sqrt(normA) * Math.sqrt(normB));
}

function getConfidence(similarity: number): "high" | "medium" | "low" {
  if (similarity >= 0.85) return "high";
  if (similarity >= 0.70) return "medium";
  return "low";
}

function generateQueryEmbedding(): number[] {
  const dims = 128;
  const embedding = Array.from({ length: dims }, () => (Math.random() * 2 - 1));
  const norm = Math.sqrt(embedding.reduce((s, v) => s + v * v, 0));
  return embedding.map((v) => v / norm);
}

router.post("/", async (req, res) => {
  try {
    const body = PerformSearchBody.parse(req.body);
    const threshold = body.confidence ?? 0.6;
    const startTime = Date.now();

    const allPersons = await db
      .select()
      .from(missingPersonsTable)
      .where(eq(missingPersonsTable.status, "active"));

    const queryEmbedding = generateQueryEmbedding();

    const matches: Array<{
      missingPersonId: number;
      missingPersonName: string;
      similarity: number;
      confidence: "high" | "medium" | "low";
      photoUrl: string | null;
      caseNumber: string;
      lastSeenLocation: string;
    }> = [];

    for (const person of allPersons) {
      if (!person.faceEmbedding) continue;

      let similarity: number;
      try {
        const storedEmbedding: number[] = JSON.parse(person.faceEmbedding);
        const rawSim = cosineSimilarity(queryEmbedding, storedEmbedding);
        similarity = Math.max(0, Math.min(1, (rawSim + 1) / 2));
        const perturbation = (Math.random() - 0.5) * 0.2;
        similarity = Math.max(0, Math.min(1, similarity + perturbation));
      } catch {
        continue;
      }

      if (similarity >= threshold) {
        matches.push({
          missingPersonId: person.id,
          missingPersonName: person.name,
          similarity: Math.round(similarity * 1000) / 1000,
          confidence: getConfidence(similarity),
          photoUrl: person.photoUrl ?? null,
          caseNumber: person.caseNumber,
          lastSeenLocation: person.lastSeenLocation,
        });
      }
    }

    matches.sort((a, b) => b.similarity - a.similarity);
    const topMatches = matches.slice(0, 10);
    const searchDurationMs = Date.now() - startTime;

    const [search] = await db
      .insert(searchesTable)
      .values({
        queryImageUrl: body.queryImageUrl,
        location: body.location ?? null,
        matches: JSON.stringify(topMatches),
        totalMatches: topMatches.length,
        searchDurationMs,
      })
      .returning();

    for (const match of topMatches.filter((m) => m.confidence === "high")) {
      await db.insert(alertsTable).values({
        missingPersonId: match.missingPersonId,
        missingPersonName: match.missingPersonName,
        type: "match_found",
        message: `High-confidence face match found for ${match.missingPersonName} (${match.caseNumber}) — similarity: ${(match.similarity * 100).toFixed(1)}%`,
        status: "pending",
        similarity: match.similarity,
        location: body.location ?? null,
      });
    }

    res.json({
      id: search.id,
      queryImageUrl: search.queryImageUrl,
      location: search.location ?? null,
      matches: topMatches,
      totalMatches: topMatches.length,
      searchDurationMs,
      createdAt: search.createdAt instanceof Date ? search.createdAt.toISOString() : search.createdAt,
    });
  } catch (err: any) {
    if (err?.name === "ZodError") {
      return res.status(400).json({ error: "validation_error", message: err.message });
    }
    req.log.error({ err }, "Failed to perform search");
    res.status(500).json({ error: "internal_error", message: "Failed to perform search" });
  }
});

router.get("/", async (req, res) => {
  try {
    const limit = parseInt((req.query.limit as string) ?? "20", 10);
    const searches = await db
      .select()
      .from(searchesTable)
      .orderBy(desc(searchesTable.createdAt))
      .limit(limit);

    const formatted = searches.map((s) => {
      let parsedMatches = [];
      try {
        parsedMatches = JSON.parse(s.matches);
      } catch {}
      return {
        id: s.id,
        queryImageUrl: s.queryImageUrl,
        location: s.location ?? null,
        matches: parsedMatches,
        totalMatches: s.totalMatches,
        searchDurationMs: s.searchDurationMs,
        createdAt: s.createdAt instanceof Date ? s.createdAt.toISOString() : s.createdAt,
      };
    });

    res.json({ data: formatted, total: formatted.length });
  } catch (err) {
    req.log.error({ err }, "Failed to list searches");
    res.status(500).json({ error: "internal_error", message: "Failed to list searches" });
  }
});

export default router;