# lib/api-zod/src/generated/types/createAlertRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TraceFaceNet Missing Person Identification API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateAlertRequestType } from "./createAlertRequestType";

export interface CreateAlertRequest {
  missingPersonId?: number | null;
  type: CreateAlertRequestType;
  message: string;
  similarity?: number | null;
  location?: string | null;
}