# lib/api-zod/src/generated/types/createMissingPersonRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TraceFaceNet Missing Person Identification API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateMissingPersonRequestGender } from "./createMissingPersonRequestGender";

export interface CreateMissingPersonRequest {
  name: string;
  age: number;
  gender: CreateMissingPersonRequestGender;
  description?: string;
  lastSeenLocation: string;
  lastSeenDate: string;
  contactName: string;
  contactPhone: string;
  contactEmail?: string;
  photoUrl?: string | null;
}