# lib/api-zod/src/generated/types/faceMatch.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TraceFaceNet Missing Person Identification API
 * OpenAPI spec version: 0.1.0
 */
import type { FaceMatchConfidence } from "./faceMatchConfidence";

export interface FaceMatch {
  missingPersonId: number;
  missingPersonName: string;
  similarity: number;
  confidence: FaceMatchConfidence;
  photoUrl?: string | null;
  caseNumber: string;
  lastSeenLocation: string;
}