# lib/api-zod/src/generated/types/missingPersonListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TraceFaceNet Missing Person Identification API
 * OpenAPI spec version: 0.1.0
 */
import type { MissingPerson } from "./missingPerson";

export interface MissingPersonListResponse {
  data: MissingPerson[];
  total: number;
  limit: number;
  offset: number;
}