# lib/api-zod/src/generated/types/searchListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TraceFaceNet Missing Person Identification API
 * OpenAPI spec version: 0.1.0
 */
import type { SearchResult } from "./searchResult";

export interface SearchListResponse {
  data: SearchResult[];
  total: number;
}