# lib/api-zod/src/generated/types/statsResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TraceFaceNet Missing Person Identification API
 * OpenAPI spec version: 0.1.0
 */

export interface StatsResponse {
  totalCases: number;
  activeCases: number;
  foundCases: number;
  closedCases: number;
  totalSearches: number;
  totalAlerts: number;
  pendingAlerts: number;
  recentMatches: number;
}